+++ sciguppy/expit.py
@@ -16,6 +16,7 @@ CACHE_DIR = os.path.join(CUR_DIR, 'cache')
 
 mod = SourceModule(open(os.path.join(CUR_DIR, 'kernel/expit.cu')).read(), cache_dir=CACHE_DIR)
 expit_kernel = mod.get_function('expit_kernel')
+expit_fast_kernel = mod.get_function('expit_fast_kernel')
 
 @gpu_func
 def expit(d_a, mode=MathModes.ACC):
@@ -26,6 +27,7 @@ def expit(d_a, mode=MathModes.ACC):
     d_out = pycuda.gpuarray.zeros(d_a.shape, numpy.float32)
     thread_size = min(d_a.size, MAX_BLOCK_SIZE)
     block_size = max(int(math.ceil(d_a.size / float(thread_size))), 1)
-    expit_kernel(d_a, d_out, numpy.int32(d_a.size), numpy.int32(mode==MathModes.FAST),
+    kernel = expit_fast_kernel if mode == MathModes.FAST else expit_kernel
+    kernel(d_a, d_out, numpy.int32(d_a.size),
             block=(thread_size,1,1), grid=(block_size,1,1))
     return d_out